+++ source/difference.js
@@ -1,0 +1,9 @@
+import { filter } from './filter.js'
+import { includes } from './includes.js'
+
+export function difference(x) {
+	return y => ([
+		...filter(value => !includes(value)(y))(x),
+		...filter(value => !includes(value)(x))(y),
+	])
+}