+++ source/exists.js
@@ -1,0 +1,7 @@
+import { find } from './find.js'
+
+export function exists(predicate) {
+  return list => {
+		return find(predicate)(list) !== undefined
+  }
+}