+++ source/intersectionWith.spec.js
@@ -1,4 +1,4 @@
-import { innerJoin } from './innerJoin.js'
+import { intersectionWith } from './intersectionWith.js'
 
 const a = {
   id: 1,
@@ -12,7 +12,7 @@ const c = {
   id: 3,
   name: 'c',
 }
-const f = (a, b) => innerJoin((r, id) => r.id === id, a)(b)
+const f = (a, b) => intersectionWith((r, id) => r.id === id, a)(b)
 
 test('only returns elements from the first list', () => {
   expect(f([a, b, c], [])).toEqual([])
@@ -31,6 +31,6 @@ test('readme example', () => {
   const list1 = [1, 2, 3, 4, 5]
   const list2 = [4, 5, 6]
   const predicate = (x, y) => x >= y
-  const result = innerJoin(predicate, list1)(list2)
+  const result = intersectionWith(predicate, list1)(list2)
   expect(result).toEqual([4, 5])
 })