+++ source/range.spec.js
@@ -1,7 +1,8 @@
 import { range } from './range.js'
 
 test('happy', () => {
-  expect(range(0)(5)).toEqual([0, 1, 2, 3, 4])
-	expect(range(7)(3)).toEqual([7, 6, 5, 4])
-	expect(range(5)(5)).toEqual([])
+  expect(range(5)).toEqual([0, 1, 2, 3, 4, 5])
+  expect(range(3,5)).toEqual([3, 4, 5])
+  expect(range(5,3)).toEqual([])
+	expect(range(0)).toEqual([])
 })